refactor revert tx and rollback tx tool
