+++ crates/rooch/src/commands/db/commands/revert_tx.rs
@@ -3,17 +3,13 @@
 
 use anyhow::Error;
 use clap::Parser;
-use metrics::RegistryService;
-use moveos_types::moveos_std::object::ObjectMeta;
-use rooch_config::{RoochOpt, R_OPT_NET_HELP};
-use rooch_db::RoochDB;
-use rooch_genesis::RoochGenesis;
+use rooch_config::R_OPT_NET_HELP;
 use rooch_types::error::{RoochError, RoochResult};
 use rooch_types::rooch_network::RoochChainID;
 use std::path::PathBuf;
-use std::time::SystemTime;
 
 use crate::cli_types::WalletContextOptions;
+use crate::commands::db::commands::init;
 
 /// Revert tx by db command.
 #[derive(Debug, Parser)]
@@ -38,7 +34,7 @@ pub struct RevertTxCommand {
 impl RevertTxCommand {
     pub async fn execute(self) -> RoochResult<()> {
         let tx_order = self.tx_order;
-        let (_root, rooch_db, _start_time) = self.init();
+        let (_root, rooch_db, _start_time) = init(self.base_data_dir, self.chain_id);
 
         let tx_hashes = rooch_db
             .rooch_store
@@ -57,17 +53,5 @@ impl RevertTxCommand {
         rooch_db.revert_tx(tx_hash)?;
 
         Ok(())
-    }
-
-    fn init(self) -> (ObjectMeta, RoochDB, SystemTime) {
-        let start_time = SystemTime::now();
-
-        let opt = RoochOpt::new_with_default(self.base_data_dir, self.chain_id, None).unwrap();
-        let registry_service = RegistryService::default();
-        let rooch_db =
-            RoochDB::init(opt.store_config(), &registry_service.default_registry()).unwrap();
-        let _genesis = RoochGenesis::load_or_init(opt.network(), &rooch_db).unwrap();
-        let root = rooch_db.latest_root().unwrap().unwrap();
-        (root, rooch_db, start_time)
     }
 }
