+++ crates/rooch/src/commands/db/commands/rollback.rs
@@ -1,32 +1,19 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use anyhow::{anyhow, Error};
+use anyhow::Error;
 use clap::Parser;
-use metrics::RegistryService;
 use moveos_store::config_store::ConfigStore;
 use moveos_store::transaction_store::TransactionStore as TxExecutionInfoStore;
-use moveos_types::access_path::AccessPath;
-use moveos_types::moveos_std::object::ObjectMeta;
 use moveos_types::startup_info;
-use moveos_types::state_resolver::{RootObjectResolver, StateReader};
-use rooch_config::{RoochOpt, R_OPT_NET_HELP};
-use rooch_db::RoochDB;
-use rooch_genesis::RoochGenesis;
-use rooch_indexer::store::traits::IndexerStoreTrait;
-use rooch_store::state_store::StateStore;
+use rooch_config::R_OPT_NET_HELP;
 use rooch_types::error::{RoochError, RoochResult};
-use rooch_types::indexer::state::{
-    collect_revert_object_change_ids, handle_revert_object_change, IndexerObjectStateChangeSet,
-    IndexerObjectStatesIndexGenerator,
-};
 use rooch_types::rooch_network::RoochChainID;
 use rooch_types::sequencer::SequencerInfo;
-use std::collections::HashMap;
 use std::path::PathBuf;
-use std::time::SystemTime;
 
 use crate::cli_types::WalletContextOptions;
+use crate::commands::db::commands::init;
 
 /// Rollback the state to a specific transaction order.
 #[derive(Debug, Parser)]
@@ -51,7 +38,7 @@ pub struct RollbackCommand {
 impl RollbackCommand {
     pub async fn execute(self) -> RoochResult<()> {
         let tx_order = self.tx_order;
-        let (_root, rooch_db, _start_time) = self.init();
+        let (_root, rooch_db, _start_time) = init(self.base_data_dir, self.chain_id);
 
         let last_sequencer_info = rooch_db
             .rooch_store
@@ -126,100 +113,7 @@ impl RollbackCommand {
                 ))));
             }
             let tx_hash = tx_hashes[0].unwrap();
-            let ledger_tx_opt = rooch_db
-                .rooch_store
-                .transaction_store
-                .get_transaction_by_hash(tx_hash)?;
-
-            if ledger_tx_opt.is_none() {
-                println!("the ledger tx not exist via tx_hash {}", tx_hash);
-                continue;
-            }
-
-            let sequencer_info = ledger_tx_opt.unwrap().sequence_info;
-            let tx_order = sequencer_info.tx_order;
-            rooch_db
-                .rooch_store
-                .transaction_store
-                .remove_transaction(tx_hash, tx_order)?;
-            rooch_db
-                .moveos_store
-                .transaction_store
-                .remove_tx_execution_info(tx_hash)?;
-            println!(
-                "remove tx succ, tx_hash: {:?}, tx_order {}",
-                tx_hash, tx_order
-            );
-
-            // remove the state change set
-            let state_change_set_ext_opt = rooch_db.rooch_store.get_state_change_set(tx_order)?;
-            if state_change_set_ext_opt.is_some() {
-                rooch_db.rooch_store.remove_state_change_set(tx_order)?;
-            }
-
-            // revert the indexer
-            let previous_tx_order = if tx_order > 0 { tx_order - 1 } else { 0 };
-            let previous_state_change_set_ext_opt = rooch_db
-                .rooch_store
-                .get_state_change_set(previous_tx_order)?;
-            if previous_state_change_set_ext_opt.is_some() && state_change_set_ext_opt.is_some() {
-                let previous_state_change_set_ext = previous_state_change_set_ext_opt.unwrap();
-                let state_change_set_ext = state_change_set_ext_opt.unwrap();
-
-                let mut object_ids = vec![];
-                for (_feild_key, object_change) in
-                    state_change_set_ext.state_change_set.changes.clone()
-                {
-                    collect_revert_object_change_ids(object_change, &mut object_ids)?;
-                }
-
-                let root = ObjectMeta::root_metadata(
-                    previous_state_change_set_ext.state_change_set.state_root,
-                    previous_state_change_set_ext.state_change_set.global_size,
-                );
-                let resolver = RootObjectResolver::new(root, &rooch_db.moveos_store);
-                let object_mapping = resolver
-                    .get_states(AccessPath::objects(object_ids))?
-                    .into_iter()
-                    .flatten()
-                    .map(|v| (v.metadata.id.clone(), v.metadata))
-                    .collect::<HashMap<_, _>>();
-
-                // 1. revert indexer transaction
-                rooch_db
-                    .indexer_store
-                    .delete_transactions(vec![tx_order])
-                    .map_err(|e| anyhow!(format!("Revert indexer transactions error: {:?}", e)))?;
-
-                // 2. revert indexer event
-                rooch_db
-                    .indexer_store
-                    .delete_events(vec![tx_order])
-                    .map_err(|e| anyhow!(format!("Revert indexer events error: {:?}", e)))?;
-
-                // 3. revert indexer full object state, including object_states, utxos and inscriptions
-                // indexer object state index generator
-                let mut state_index_generator = IndexerObjectStatesIndexGenerator::default();
-                let mut indexer_object_state_change_set = IndexerObjectStateChangeSet::default();
-
-                for (_feild_key, object_change) in state_change_set_ext.state_change_set.changes {
-                    handle_revert_object_change(
-                        &mut state_index_generator,
-                        tx_order,
-                        &mut indexer_object_state_change_set,
-                        object_change,
-                        &object_mapping,
-                    )?;
-                }
-                rooch_db
-                    .indexer_store
-                    .apply_object_states(indexer_object_state_change_set)
-                    .map_err(|e| anyhow!(format!("Revert indexer states error: {:?}", e)))?;
-            }
-            println!(
-                "revert indexer succ, tx_hash: {:?}, tx_order {}",
-                tx_hash, tx_order
-            );
+            rooch_db.do_revert_tx_ignore_check(tx_hash)?;
         }
 
         let rollback_sequencer_info = SequencerInfo {
@@ -240,17 +134,5 @@ impl RollbackCommand {
             tx_hash, tx_order, execution_info.state_root
         );
         Ok(())
-    }
-
-    fn init(self) -> (ObjectMeta, RoochDB, SystemTime) {
-        let start_time = SystemTime::now();
-
-        let opt = RoochOpt::new_with_default(self.base_data_dir, self.chain_id, None).unwrap();
-        let registry_service = RegistryService::default();
-        let rooch_db =
-            RoochDB::init(opt.store_config(), &registry_service.default_registry()).unwrap();
-        let _genesis = RoochGenesis::load_or_init(opt.network(), &rooch_db).unwrap();
-        let root = rooch_db.latest_root().unwrap().unwrap();
-        (root, rooch_db, start_time)
     }
 }
